requirements.txt
amqp<=5.1.1
asgiref<=3.5.2
async-timeout<=4.0.2
beautifulsoup4<=4.11.1
billiard<=3.6.4.0
boto3<=1.24.68
botocore<=1.27.68
celery==5.2.7
click<=8.1.3
click-didyoumean<=0.3.0
click-plugins<=1.1.1
click-repl<=0.2.0
colorama<=0.4.5
Deprecated<=1.2.13
Django<=4.0.8
django-celery-beat==2.3.0
django-storages<=1.13.1
django-timezone-field<=5.0
flower<=1.2.0
gunicorn<=20.1.0
humanize<=4.4.0
jmespath<=1.0.1
kombu<=5.2.4
mysql-connector-python<=8.0.30
mysqlclient<=2.1.1
packaging<=21.3
prometheus-client<=0.14.1
prompt-toolkit<=3.0.31
protobuf<=3.20.1
pyparsing<=3.0.9
python-crontab<=2.6.0
python-dateutil<=2.8.2
pytz<=2022.4
redis<=4.3.4
s3transfer<=0.6.0
six<=1.16.0
soupsieve<=2.3.2.post1
sqlparse<=0.4.2
tornado<=6.2
tzdata<=2022.2
urllib3<=1.26.12
vine<=5.0.0
wcwidth<=0.2.5
whitenoise<=6.2.0
wrapt<=1.14.1